+++ src/moonblade/mod.rs
@@ -12,6 +12,6 @@ mod utils;
 pub use self::agg::{AggregationProgram, GroupAggregationProgram, Stats};
 pub use self::choose::ChooseProgram;
 pub use self::error::{ConcretizationError, EvaluationError, SpecifiedEvaluationError};
-pub use self::interpreter::Program;
+pub use self::interpreter::{GlobalVariables, Program};
 pub use self::select::SelectionProgram;
 pub use self::types::DynamicValue;